[server] | refactor(webhook-transcoder): field명 변경
