+++ packages/server/src/webhook/transcoder/transcoder-webhook.controller.ts
@@ -4,7 +4,7 @@ import { TranscoderNotificationState } from 'src/constants';
 import { Repository } from 'typeorm';
 import { Video } from '../../../../typeorm/src/entity/video.entity';
 import { Tag } from '../../../../typeorm/src/entity/tag.entity';
-import { UploadedVideoTableService } from 'src/uploaded-video/uploaded-video-table.service';
+import { UploadedVideoTableService } from 'src/uploaded-video-table/uploaded-video-table.service';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from '../../../../typeorm/src/entity/user.entity';
 
@@ -31,7 +31,7 @@ export class TranscoderWebhookController {
       const id = path[path.length - 2];
 
       const videoInfo = this.uploadedVideoTableService.find(id);
-      videoInfo.sourceUrl = mpdSourceUrl;
+      videoInfo.source = mpdSourceUrl;
 
       const user = await this.userRepository.find({ where: { id: 0 } });
 
@@ -61,7 +61,7 @@ export class TranscoderWebhookController {
       const video = this.videoRepository.create({
         title: videoInfo.title,
         description: videoInfo.description,
-        sourceUrl: videoInfo.sourceUrl,
+        sourceUrl: videoInfo.source,
         thumbnail: 'thumbnail',
         playtime: notificationMessage.outputs[0].duration,
         tags,
